--- src/main/java/com/tropical/backend/smalltalk/service/AIGenerationTopicService.java
@@ -1,7 +1,0 @@
-package com.tropical.backend.smalltalk.service;
-
-import org.springframework.stereotype.Service;
-
-@Service
-public class AIGenerationTopicService {
-}
